Move the C11 injectable-git test into its own process
The test proving SGT_GIT_BIN redirects runtime::git's invocations lived
in the lib test binary and mutated the process-global environment around
a live git() call. std::env::set_var is process-global, not
thread-scoped, and command() reads SGT_GIT_BIN fresh per invocation (by
design), so for the duration of the override every other test thread
that shelled out through this module ran the scripted stub instead of
git. The SAFETY comment's claim that "the window is exactly one
command() call" bounded the window in time but not across threads.

Measured at the previous commit: running the lib test binary 60 times as
`runtime::git::tests --test-threads=8` failed 38 times, always in the
unrelated, unmodified failure_carries_gits_own_diagnostic, in three
shapes -- the stub's stdout arriving where git's "not a git repository"
diagnostic was expected, rev-parse "succeeding" in a non-repository
because the stub always exits zero, and the override outliving the
TempDir holding the script (`cannot open .../scripted-git`). The third
shape shows the leak can outlast the override's own tempdir, so every
call site reaching real Git through this module was exposed, not just
that one test.

Fixed by moving the test to tests/c11_injectable_git.rs. Cargo gives
every integration-test file its own process, so a file holding exactly
one test has no sibling thread to leak into -- and the end-to-end proof
is unchanged, since runtime::git is public. The file header and the
GIT_BIN_ENV doc comment both record that this file must stay a
single-test file and why; a second test added there reintroduces the
race. No production code changes: command()'s per-call read is the
pinned C11 behavior and stays exactly as it was.

Same 60-run stress after the move: 0 failures.
